refactoring Bind and tests
