+++ tests/database/DataMapper/Statement/Bind/InlineTest.php
@@ -15,10 +15,14 @@ namespace Phalcon\Tests\Database\DataMapper\Statement\Bind;
 
 use PDO;
 use Phalcon\DataMapper\Statement\Bind;
+use Phalcon\DataMapper\Statement\Select;
 use Phalcon\Tests\AbstractDatabaseTestCase;
+
 use ReflectionClass;
 
-final class CloneTest extends AbstractDatabaseTestCase
+use function env;
+
+final class InlineTest extends AbstractDatabaseTestCase
 {
     /**
      * @return void
@@ -39,21 +43,28 @@ final class CloneTest extends AbstractDatabaseTestCase
     }
 
     /**
-     * Database Tests Phalcon\DataMapper\Statement\Bind :: clone()
+     * Database Tests Phalcon\DataMapper\Statement\Bind :: bindInline()
      *
      * @since  2020-01-20
      *
      * @group  common
      */
-    public function testDmQueryBindClone(): void
+    public function testDmQueryBindBindInline(): void
     {
-        $bind = new Bind();
+        $bind   = new Bind();
 
         $expected = [];
         $actual   = $bind->toArray();
         $this->assertSame($expected, $actual);
 
         $bind->inline("one");
+
+        $expected = [
+            "_1_1_" => ["one", 2],
+        ];
+        $actual   = $bind->toArray();
+        $this->assertSame($expected, $actual);
+
         $bind->inline(true, PDO::PARAM_BOOL);
 
         $expected = [
@@ -63,16 +74,29 @@ final class CloneTest extends AbstractDatabaseTestCase
         $actual   = $bind->toArray();
         $this->assertSame($expected, $actual);
 
-        $clone = clone $bind;
-
-        $clone->inline("two");
+        $bind->inline(["six", "seven", 8, 9]);
 
         $expected = [
             "_1_1_" => ["one", 2],
             "_1_2_" => [true, 5],
-            "_2_3_" => ["two", 2],
+            "_1_3_" => ["six", 2],
+            "_1_4_" => ["seven", 2],
+            "_1_5_" => [8, 1],
+            "_1_6_" => [9, 1],
         ];
-        $actual   = $clone->toArray();
+        $actual   = $bind->toArray();
+        $this->assertSame($expected, $actual);
+
+        $driver = env('driver');
+        $select = Select::new($driver);
+
+        $select
+            ->from('co_customers')
+            ->where('inv_cst_id = ', 1)
+        ;
+
+        $expected = '(SELECT * FROM co_customers WHERE inv_cst_id = :_2_1_)';
+        $actual   = $bind->inline($select);
         $this->assertSame($expected, $actual);
     }
 }
