+++ tests/database/DataMapper/Statement/Bind/MergeRemoveResetTest.php
@@ -13,11 +13,12 @@ declare(strict_types=1);
 
 namespace Phalcon\Tests\Database\DataMapper\Statement\Bind;
 
+use PDO;
 use Phalcon\DataMapper\Statement\Bind;
 use Phalcon\Tests\AbstractDatabaseTestCase;
 use ReflectionClass;
 
-final class SetValuesTest extends AbstractDatabaseTestCase
+final class MergeRemoveResetTest extends AbstractDatabaseTestCase
 {
     /**
      * @return void
@@ -38,13 +39,13 @@ final class SetValuesTest extends AbstractDatabaseTestCase
     }
 
     /**
-     * Database Tests Phalcon\DataMapper\Statement\Bind :: setValues()
+     * Database Tests Phalcon\DataMapper\Statement\Bind :: merge()/remove()/reset()
      *
      * @since  2020-01-20
      *
      * @group  common
      */
-    public function testDmQueryBindSetValues(): void
+    public function testDmQueryBindMergeRemoveReset(): void
     {
         $bind = new Bind();
 
@@ -52,26 +53,44 @@ final class SetValuesTest extends AbstractDatabaseTestCase
         $actual   = $bind->toArray();
         $this->assertSame($expected, $actual);
 
-        $bind->setValues(
-            [
-                "one"   => "two",
-                "three" => "four",
-                "five"  => ["six", "seven", 8, 9],
-            ]
-        );
-
-        $bind->setValue("nine", null);
-        $bind->setValue("ten", false);
-        $bind->setValue("eleven", 11);
+        $bind->inline("one");
+        $bind->inline(true, PDO::PARAM_BOOL);
 
         $expected = [
-            "one"    => ["two", 2],
-            "three"  => ["four", 2],
-            "five"   => [['six', 'seven', 8, 9], 2],
-            "nine"   => [null, 0],
-            "ten"    => [false, 5],
-            "eleven" => [11, 1],
+            "_1_1_" => ["one", 2],
+            "_1_2_" => [true, 5],
         ];
+        $actual   = $bind->toArray();
+        $this->assertSame($expected, $actual);
+
+        $values = [
+            "three" => "four",
+            "five"  => ["six", "seven", 8, 9],
+        ];
+        $bind->merge($values);
+
+        $expected = [
+            "_1_1_" => ["one", 2],
+            "_1_2_" => [true, 5],
+            "three" => "four",
+            "five"  => ["six", "seven", 8, 9],
+        ];
+        $actual   = $bind->toArray();
+        $this->assertSame($expected, $actual);
+
+        $bind->remove("_1_1_");
+
+        $expected = [
+            "_1_2_" => [true, 5],
+            "three" => "four",
+            "five"  => ["six", "seven", 8, 9],
+        ];
+        $actual   = $bind->toArray();
+        $this->assertSame($expected, $actual);
+
+        $bind->reset();
+
+        $expected = [];
         $actual   = $bind->toArray();
         $this->assertSame($expected, $actual);
     }
